WIP: Fill the endpoints of the comment controller and update some endpoint for the article.

+++ src/main/java/de/frauas/userauth/service/ArticleService.java
@@ -2,10 +2,14 @@ package de.frauas.userauth.service;
 
 import de.frauas.userauth.dto.ArticleDto;
 import de.frauas.userauth.entity.Article;
+import de.frauas.userauth.entity.Comment;
 import de.frauas.userauth.entity.User;
 import de.frauas.userauth.exceptions.ArticleNotFoundException;
+import de.frauas.userauth.exceptions.CommentNotFoundException;
+import de.frauas.userauth.exceptions.UnauthorizedException;
 import de.frauas.userauth.repository.ArticleRepository;
-import de.frauas.userauth.util.SecurityUtil;
+import de.frauas.userauth.repository.CommentRepository;
+import de.frauas.userauth.util.JwtTokenUtil;
 import org.springframework.stereotype.Service;
 
 import java.util.Collections;
@@ -16,12 +20,15 @@ import java.util.Optional;
 public class ArticleService {
 
     private final ArticleRepository articleRepository;
+    private final CommentRepository commentRepository;
+    private final JwtTokenUtil jwtTokenUtil;
+    private final UserService userService;
 
-    private final SecurityUtil securityUtil;
-
-    public ArticleService(ArticleRepository articleRepository, SecurityUtil securityUtil) {
+    public ArticleService(ArticleRepository articleRepository, CommentRepository commentRepository, JwtTokenUtil jwtTokenUtil, UserService userService) {
         this.articleRepository = articleRepository;
-        this.securityUtil = securityUtil;
+        this.commentRepository = commentRepository;
+        this.jwtTokenUtil = jwtTokenUtil;
+        this.userService = userService;
     }
 
     public Optional<Article> findArticleById(Long id) {
@@ -32,7 +39,8 @@ public class ArticleService {
         return articleRepository.findAll();
     }
 
-    public void createArticle(ArticleDto articleDto, User author) {
+    public void createArticle(ArticleDto articleDto, String header) {
+        User author = userService.findUserByUserName(jwtTokenUtil.getUsernameFromToken(jwtTokenUtil.getTokenFromAuthorizationHeader(header)));
         Article article = Article.builder()
                 .headline(articleDto.getHeadline())
                 .content(articleDto.getContent())
@@ -43,15 +51,67 @@ public class ArticleService {
         articleRepository.save(article);
     }
 
-    public void deleteArticle(Long id) {
+    public void deleteArticle(Long id, String header) {
+        String userName = jwtTokenUtil.getUsernameFromToken(jwtTokenUtil.getTokenFromAuthorizationHeader(header));
+        Article existingArticle = findArticleById(id).orElseThrow(ArticleNotFoundException::new);
+
+        if (!existingArticle.getAuthor().getUsername().equals(userName)) {
+            throw new UnauthorizedException();
+        }
+
         articleRepository.deleteById(id);
     }
 
-    public void updateArticle(Long id, String headline, String content) {
-        Article existingArticle = findArticleById(id).orElseThrow(ArticleNotFoundException::new);
-        existingArticle.setHeadline(headline);
-        existingArticle.setContent(content);
-        articleRepository.save(existingArticle);
+    public void updateArticle(ArticleDto articleDto, String header) {
+        String userName = jwtTokenUtil.getUsernameFromToken(jwtTokenUtil.getTokenFromAuthorizationHeader(header));
+        Article existingArticle = findArticleById(articleDto.getId()).orElseThrow(ArticleNotFoundException::new);
 
+        if (!existingArticle.getAuthor().getUsername().equals(userName)) {
+            throw new UnauthorizedException();
+        }
+
+        existingArticle.setHeadline(articleDto.getHeadline());
+        existingArticle.setContent(articleDto.getContent());
+        articleRepository.save(existingArticle);
+    }
+
+    public void addComment(Long articleId, String comment, String header) {
+        User author = userService.findUserByUserName(jwtTokenUtil.getUsernameFromToken(jwtTokenUtil.getTokenFromAuthorizationHeader(header)));
+        Article existingArticle = findArticleById(articleId).orElseThrow(ArticleNotFoundException::new);
+
+        List<Comment> existingComments = existingArticle.getComments();
+        existingComments.add(Comment.builder().content(comment).author(author).build());
+        existingArticle.setComments(existingComments);
+
+        articleRepository.save(existingArticle);
+    }
+
+    public Optional<Comment> findCommentById(Long id) {
+        return commentRepository.findById(id);
+    }
+
+    public void deleteComment(Long id, String header) {
+        String userName = jwtTokenUtil.getUsernameFromToken(jwtTokenUtil.getTokenFromAuthorizationHeader(header));
+        Comment existingComment = findCommentById(id).orElseThrow(CommentNotFoundException::new);
+
+        if (!existingComment.getAuthor().getUsername().equals(userName)) {
+            throw new UnauthorizedException();
+        }
+
+        commentRepository.deleteById(id);
+
+    }
+
+    public void updateComment(Long id, String content, String header) {
+        String userName = jwtTokenUtil.getUsernameFromToken(jwtTokenUtil.getTokenFromAuthorizationHeader(header));
+        Comment existingComment = findCommentById(id).orElseThrow(CommentNotFoundException::new);
+
+        if (!existingComment.getAuthor().getUsername().equals(userName)) {
+            throw new UnauthorizedException();
+        }
+
+        existingComment.setContent(content);
+
+        commentRepository.save(existingComment);
     }
 }
